Move Old/NG common functions in common_fixture
