+++ artemis/tests/bibus_test.py
@@ -1,6 +1,6 @@
 from artemis import default_checker
 from artemis.default_checker import stop_schedule_checker
-from artemis.test_mechanism import dataset, DataSet, set_scenario
+from artemis.common_fixture import dataset, DataSet, set_scenario
 from artemis.tests.fixture import ArtemisTestFixture
 import pytest
